Extract functions config to avoid dependency cycles, assign failure policy to triggers
