+++ src/providers/https.ts
@@ -26,15 +26,12 @@ import * as _ from 'lodash';
 import * as cors from 'cors';
 import { apps } from '../apps';
 import { HttpsFunction, optionsToTrigger, Runnable } from '../cloud-functions';
-import { DeploymentOptions } from '../function-builder';
+import { DeploymentOptions } from '../function-configuration';
 
-/**
- *
- *
- */
 export interface Request extends express.Request {
   rawBody: Buffer;
 }
+
 /**
  * Handle HTTP requests.
  * @param handler A function that takes a request and response object,
